[Bug] Code modification done (#301)
1. Added check for checking if partition folder exists, then only go for deleting the folder
2. Modified message and method name to put data row in bad records
3. Trim each column data for processing as there is bug in spark csv reader which trims the white spaces even if property is set to ignore them

+++ integration/spark/src/main/scala/org/apache/spark/sql/cubemodel/cubeSchema.scala
@@ -1970,11 +1970,14 @@ private[sql] case class LoadCube(
           sys.error("Dataload failure. Please check the logs")
       }
       finally {
-        // Once the data load is successfule delete the unwanted partition files
+        // Once the data load is successful delete the unwanted partition files
         try {
-          val file = FileFactory
-            .getCarbonFile(partitionLocation, FileFactory.getFileType(partitionLocation))
-          CarbonUtil.deleteFoldersAndFiles(file)
+          val fileType = FileFactory.getFileType(partitionLocation)
+          if(FileFactory.isFileExist(partitionLocation, fileType)) {
+            val file = FileFactory
+              .getCarbonFile(partitionLocation, fileType)
+            CarbonUtil.deleteFoldersAndFiles(file)
+          }
         } catch {
           case ex: Exception =>
             LOGGER.error(CarbonSparkInterFaceLogEvent.UNIBI_CARBON_SPARK_INTERFACE_MSG, ex)
